requirements.txt
Flask
random
sklearn.preprocessing
sklearn.metrics
sklearn.model_selection
xgboost
pandas